config/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
h3==3.7.6
idna==3.4
iso3166==2.1.1
jsonpickle==3.0.1
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.2
psycopg2-binary==2.9.5
pycparser==2.21
pyrus-api==2.13.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
timezonefinder==6.2.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.8.2
